+++ TomasGreen.Web/Areas/Import/Controllers/PurchasedArticleWarehousesController.cs
@@ -28,7 +28,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         }
 
         // GET: Stock/PurchasedArticleWarehouses/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -74,7 +74,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         }
 
         // GET: Stock/PurchasedArticleWarehouses/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -96,7 +96,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("PurchasedArticleID,WarehouseID,QtyBoxes,QtyExtraKg")] PurchasedArticleWarehouse PurchasedArticleWarehouse)
+        public async Task<IActionResult> Edit(int id, [Bind("PurchasedArticleID,WarehouseID,QtyBoxes,QtyExtraKg")] PurchasedArticleWarehouse PurchasedArticleWarehouse)
         {
             if (id != PurchasedArticleWarehouse.ID)
             {
@@ -129,7 +129,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         }
 
         // GET: Stock/PurchasedArticleWarehouses/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -151,7 +151,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         // POST: Stock/PurchasedArticleWarehouses/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var PurchasedArticleWarehouse = await _context.PurchasedArticleWarehouses.SingleOrDefaultAsync(m => m.ID == id);
             _context.PurchasedArticleWarehouses.Remove(PurchasedArticleWarehouse);
@@ -159,7 +159,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool PurchasedArticleWarehouseExists(long id)
+        private bool PurchasedArticleWarehouseExists(int id)
         {
             return _context.PurchasedArticleWarehouses.Any(e => e.ID == id);
         }
