+++ TomasGreen.Web/Areas/Import/Controllers/WarehousesController.cs
@@ -30,7 +30,7 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         }
 
         // GET: Stock/Warehouses/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -105,7 +105,7 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         }
 
         // GET: Stock/Warehouses/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -126,7 +126,7 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("ID,Name,Description,Address,Phone,AddedDate,IsOnTheWay,IsCustomers,CompanySectionID")] Warehouse warehouse)
+        public async Task<IActionResult> Edit(int id, [Bind("ID,Name,Description,Address,Phone,AddedDate,IsOnTheWay,IsCustomers,CompanySectionID")] Warehouse warehouse)
         {
             if (id != warehouse.ID)
             {
@@ -196,7 +196,7 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         }
 
         // GET: Stock/Warehouses/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -216,7 +216,7 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         // POST: Stock/Warehouses/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var warehouse = await _context.Warehouses.SingleOrDefaultAsync(m => m.ID == id);
             if(warehouse != null)
@@ -228,7 +228,7 @@ namespace TomasGreen.Web.Areas.Import.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool WarehouseExists(long id)
+        private bool WarehouseExists(int id)
         {
             return _context.Warehouses.Any(e => e.ID == id);
         }
