+++ TomasGreen.Web/Areas/Import/Stock/Controllers/OrderTransportsController.cs
@@ -27,7 +27,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         }
 
         // GET: Stock/OrderTransports/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -75,7 +75,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         }
 
         // GET: Stock/OrderTransports/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -95,7 +95,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("ID, Name, AddedDate")] OrderTransport orderTransport)
+        public async Task<IActionResult> Edit(int id, [Bind("ID, Name, AddedDate")] OrderTransport orderTransport)
         {
             if (id != orderTransport.ID)
             {
@@ -135,7 +135,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         }
 
         // GET: Stock/OrderTransports/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -155,7 +155,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
         // POST: Stock/OrderTransports/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var orderTransport = await _context.OrderTranports.SingleOrDefaultAsync(m => m.ID == id);
             _context.OrderTranports.Remove(orderTransport);
@@ -163,7 +163,7 @@ namespace TomasGreen.Web.Areas.Stock.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool OrderTransportExists(long id)
+        private bool OrderTransportExists(int id)
         {
             return _context.OrderTranports.Any(e => e.ID == id);
         }
