+++ TomasGreen.Web/Areas/Packing/Controllers/PackingPlansController.cs
@@ -38,7 +38,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
         }
 
         // GET: Packing/PackingPlans/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -55,7 +55,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
             return View(packingPlan);
         }
         // GET: Packing/PackingPlans/Details/5
-        public async Task<IActionResult> EditPackingPlanMix(long? id)
+        public async Task<IActionResult> EditPackingPlanMix(int? id)
         {
             if (id == null)
             {
@@ -206,7 +206,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
         }
 
         // GET: Packing/PackingPlans/Edit/5
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -226,7 +226,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("Date,ManagerID,CompanyID")] PackingPlan packingPlan)
+        public async Task<IActionResult> Edit(int id, [Bind("Date,ManagerID,CompanyID")] PackingPlan packingPlan)
         {
             if (id != packingPlan.ID)
             {
@@ -257,7 +257,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
         }
 
         // GET: Packing/PackingPlans/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -277,7 +277,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
         // POST: Packing/PackingPlans/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(long id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var packingPlan = await _context.PackingPlans.SingleOrDefaultAsync(m => m.ID == id);
           //  _context.PackingPlans.Remove(packingPlan);
@@ -285,7 +285,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool PackingPlanExists(long id)
+        private bool PackingPlanExists(int id)
         {
             return _context.PackingPlans.Any(e => e.ID == id);
         }
@@ -369,7 +369,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
             }
         }
 
-        private decimal GetTotalWeight(long articleID, int qtyPackages, decimal qtExtra)
+        private decimal GetTotalWeight(int articleID, int qtyPackages, decimal qtExtra)
         {
             decimal result = 0;
             var article = _context.Articles.Include(u => u.ArticleUnit).Where(a => a.ID == articleID).FirstOrDefault();
@@ -388,7 +388,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
 
             return result;
         }
-        public List<Warehouse> GetFromWarehouseList(Int64 companyID)
+        public List<Warehouse> GetFromWarehouseList(int companyID)
         {
             var warehouses = (from aw in _context.ArticleWarehouseBalances
                               where aw.CompanyID == companyID && (aw.QtyPackagesOnhand > 0 || aw.QtyExtraOnhand > 0)
@@ -400,7 +400,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
 
             return warehouses;
         }
-        public List<Warehouse> GetToWarehouseList(Int64 companyID)
+        public List<Warehouse> GetToWarehouseList(int companyID)
         {
             List<Warehouse> warehouses = new List<Warehouse>();
             var company = _context.Companies.Where(c => c.ID == companyID).FirstOrDefault();
@@ -419,7 +419,7 @@ namespace TomasGreen.Web.Areas.Packing.Controllers
             }
             return warehouses;
         }
-        public List<ListItem> GetArticleList(Int64 warehouseID, Int64 companyID)
+        public List<ListItem> GetArticleList(int warehouseID, int companyID)
         {
             List<ListItem> listItems = new List<ListItem>();
             var articles = (from aw in _context.ArticleWarehouseBalances
